app/backend/requirements.txt
absl-py==1.3.0
accelerate==0.13.2
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.3.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
einops==0.5.0
einops-exts==0.0.3
ema-pytorch==0.0.10
fastapi==0.85.1
filelock==3.8.0
flake8==5.0.4
frozenlist==1.3.1
fsspec==2022.10.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h11==0.14.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.2
imagen-pytorch==1.11.15
importlib-metadata==5.0.0
joblib==1.2.0
kornia==0.6.8
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
networkx==2.8.7
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
pipenv==2022.10.12
platformdirs==2.5.2
protobuf==3.19.6
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
pyDeprecate==0.3.2
pyflakes==2.5.0
pyparsing==3.0.9
pytorch-lightning==1.7.7
pytorch-warmup==0.1.0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.1
torch==1.12.1
torchmetrics==0.10.0
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
